requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
openai>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
transformers>=4.30.0
pymupdf>=1.23.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
tqdm>=4.65.0
